Rename package regionspn to package regionlock
